Interiors are now all public
This not only shortens the amount of time it takes to create an interior, but also simplifies it as all interiors will be treated as public for now. Changes can be made in favour of factions at a later time when they start to come into play.

I have also tidied up some variables and removed all usages of the pVar system which is slower than just using normal variables.