Inventory items are now read and written from the database
Inventory items can now be created in game as an admin, and they are read and written from the database.

Player inventories are now loaded from ini files within the scriptfiles directory (in a directory called character_inventory).

This has been done because this way, the player's inventory does not need updating everytime we add new items.

Loot tables will be making their way towards the database next I think so we can dynamically alter loot tables or even add new ones to then be used with /scavcreate as well.